+++ press/tmp/trans.c
@@ -2,10 +2,10 @@
 #include <stdint.h>
 #include "trans.h"
 
-int16_t *shift_x(int16_t x, const int16_t *in, uint64_t nin)
+int16_t *shift_x(int16_t x, const int16_t *in, uint32_t nin)
 {
 	int16_t *out;
-	uint64_t i;
+	uint32_t i;
 
 	out = malloc(nin * sizeof *out);
 
@@ -16,9 +16,9 @@ int16_t *shift_x(int16_t x, const int16_t *in, uint64_t nin)
 	return out;
 }
 
-void shift_x_inplace(int16_t x, int16_t *in, uint64_t nin)
+void shift_x_inplace(int16_t x, int16_t *in, uint32_t nin)
 {
-	uint64_t i;
+	uint32_t i;
 
 	for (i = 0; i < nin; i++) {
 		in[i] += x;
@@ -35,28 +35,28 @@ static inline uint16_t unzigzag_one(int16_t x)
 	return (x >> 1) ^ -(x & 1);
 }
 
-void zigzag_inplace(int16_t *in, uint64_t nin)
+void zigzag_inplace(int16_t *in, uint32_t nin)
 {
-	uint64_t i;
+	uint32_t i;
 
 	for (i = 0; i < nin; i++) {
 		in[i] = zigzag_one(in[i]);
 	}
 }
 
-void unzigzag_inplace(int16_t *in, uint64_t nin)
+void unzigzag_inplace(int16_t *in, uint32_t nin)
 {
-	uint64_t i;
+	uint32_t i;
 
 	for (i = 0; i < nin; i++) {
 		in[i] = unzigzag_one(in[i]);
 	}
 }
 
-int16_t *zigdelta(const int16_t *in, uint64_t nin, uint64_t *nout)
+int16_t *zigdelta(const int16_t *in, uint32_t nin, uint32_t *nout)
 {
 	int16_t *out;
-	uint64_t i;
+	uint32_t i;
 
 	out = malloc(nin * sizeof *out);
 
@@ -69,9 +69,9 @@ int16_t *zigdelta(const int16_t *in, uint64_t nin, uint64_t *nout)
 }
 
 /* in[0] must be the first signal */
-void unzigdelta_inplace(int16_t *in, uint64_t nin)
+void unzigdelta_inplace(int16_t *in, uint32_t nin)
 {
-	uint64_t i;
+	uint32_t i;
 
 	for (i = 1; i < nin; i++) {
 		in[i] = in[i - 1] + unzigzag_one(in[i]);
